Rename MessageHandler into MessageHandlerBuilder and split its main method
