--- tests/unit_tests/test_MessageHandler.py
@@ -1,16 +1,0 @@
-from src.MessageHandler import MessageHandler
-
-
-def test_given_unknown_command__empty_list_is_returned():
-    assert MessageHandler({}).handle_message('!unknown') == []
-
-
-def test_given_known_command__command_result_is_returned():
-    assert MessageHandler({}).handle_message('!nyan') == ['~=[,,_,,]:3']
-
-
-def test_default_rate_limit_option_is_acted_upon():
-    message_handler = MessageHandler({'default_rate_limit': "0"})
-
-    message_handler.handle_message('!nyan')
-    assert message_handler.handle_message('!nyan') == ['~=[,,_,,]:3']
